+++ frontend/src/types/index.ts
@@ -1,4 +1,4 @@
-export type { SourceCodes, Tag, Category, Template, TemplateListItem } from './template';
+export type { SourceCodes, Tag, Category, Template, TemplateListItem, TemplateVisibility } from './template';
 export type {
   SortingKey,
   SortingOption,
